+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 x.y.z / 2015-xx-yy
 ------------------
 - added `marketInfo()` See: https://github.com/jprichardson/shapeshift.io/pull/1
+- moved testing from mocha to ava
 
 1.0.0 / 2015-04-21
 ------------------